% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_search.R
\name{gene_comb}
\alias{gene_comb}
\title{Search for gene combinations of size *n_genes* without significant effects of factor variables}
\usage{
gene_comb(data, subject = 1, timepoint = 2, condition = 3, target = 4,
  expression = 5, n_genes = 2)
}
\arguments{
\item{data}{A data frame with subject,timepoint,condition,target and expression. Expression values should be log transformed linearized Cq values.}

\item{subject}{numeric column id for subject identifier}

\item{timepoint}{numeric column id for timepoint identifier}

\item{condition}{numeric column identifier for condition identifier}

\item{target}{numeric column identifier for target/gene identifier}

\item{expression}{numeric column identifier for expression data}

\item{n_genes}{numeric, specifies how many genes should be combined}
}
\description{
Search for gene combinations of size *n_genes* without significant effects of factor variables
}
