% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generefer_plot.R
\name{generefer_plot}
\alias{generefer_plot}
\title{Plot of a generefer data frame}
\usage{
generefer_plot(dat, icc.threshold = 0.5)
}
\arguments{
\item{dat}{A data.frame() from lmer_stability()}

\item{icc.threshold}{Exclude observations based on intra-class correlation lower than threshold, defaults to 0.5}
}
\value{
A list containing a plot and a tidy sorted data frame
}
\description{
Plot of a generefer data frame
}
