% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genorm.R
\name{genorm}
\alias{genorm}
\title{Analyze reference gene stability with geNorm}
\usage{
genorm(data, sample = 1, detector = 2, cq = 3, ngenes = 2)
}
\arguments{
\item{data}{a data.frame() with sample id, target id and raw cq values. If duplicates, average cq values per sample.}

\item{sample}{column number of sample ids}

\item{detector}{column number of target ids}

\item{cq}{column number of cq values}
}
\value{
A list from NormqPCR with genorm parameters
}
\description{
Analyze reference gene stability with geNorm
}
