% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfinder.R
\name{normfinder}
\alias{normfinder}
\title{Analyze a data set for gene expression stability with the NormFinder algorithm}
\usage{
normfinder(data, Groups = TRUE, ctVal = TRUE, pStabLim = 0.25,
  sample.cl = 1, gene.cl = 2, groups.cl = 3, cq.cl = 4)
}
\arguments{
\item{data}{A data frame with sample,gene and group identifier and expression values. Expression values can be raw Cq or on linear scale.}

\item{Groups}{Logical, if TRUE, the function will consider grouping variables}

\item{ctVal}{Logical, TRUE if raw Cq values}

\item{pStabLim, }{numeric value on stability threshold.}

\item{sample.cl}{numeric specifies the column of sample id}

\item{gene.cl}{numeric specifies the column of gene id}

\item{groups.cl}{numeric specifies the column of groups id}

\item{cq.cl}{numeric specifies the column of cq/expression values}
}
\description{
Analyze a data set for gene expression stability with the NormFinder algorithm
}
