% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_genorm.R
\name{tidy_genorm}
\alias{tidy_genorm}
\title{Calculate reference gene stability}
\usage{
tidy_genorm(dat, sample = "sample", gene = "gene",
  expression = "expression", log = FALSE)
}
\arguments{
\item{dat}{Name of the data frame, supplied to the function in i tidy way. 
A column for sample id ("sample"), a column for gene id ("gene") and a column 
containing expression values ("expression").}

\item{sample}{Character column name of column containing sample id's}

\item{gene}{Character column name of column containing gene id's}

\item{expression}{Character column name of column containing sample expression values}

\item{log}{Logical, is data on the log scale?}
}
\value{
A list containing: M, a data frame with stability measures, average stability measures per gene after step-wise
exclusion of genes; Normalisation factors for n >= 2 genes per sample, and; Pairwise variations of normalisation factors.
}
\description{
Calculates M values for n >= 2 genes. When there are more than two genes in the data set, 
the function calculates M for the full data set, removes the least stable gene (largest M) and 
calculates M again until only two genes remains. The remaining two genes cannot be ranked. 
Based on gene ranking normalisation factors (NF) are calculated based on two to n genes and pairwise
variations between NFn and NFn+1 are calculated which can be used to determine the optimal
number of reference genes to include in the experiment.
}
\references{
Vandesompele, J., et al. (2002). "Accurate normalization of real-time quantitative 
RT-PCR data by geometric averaging of multiple internal control genes." Genome Biol 3(7)
}
