% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_normfinder.R
\name{tidy_normfinder}
\alias{tidy_normfinder}
\title{Analyze a data set for gene expression stability with the NormFinder algorithm}
\usage{
tidy_normfinder(data, group = TRUE, sample = "sample", gene = "gene",
  groups = "group", expression = "expression", log = FALSE)
}
\arguments{
\item{data}{A data frame with sample, gene and group (optional) identifier and expression values. 
Expression values can be raw Cq or on linear scale.}

\item{group}{Logical, if TRUE, the function will consider a grouping variable}

\item{sample}{Character Column name for the sample identifiers}

\item{gene}{Character Column name of the gene identifiers}

\item{groups}{Character Column name of the group identifiers}

\item{expression}{Character Name of column containing expression values}

\item{log}{Logical if TRUE the function will consider expression values to be on a log scale}
}
\description{
Analyze a data set for gene expression stability with the NormFinder algorithm
}
