% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfinder.R
\name{normfinder}
\alias{normfinder}
\title{Analyze a data set for gene expression stability with the NormFinder algorithm}
\usage{
normfinder(data, group = TRUE, ctVal = FALSE, pStabLim = 0.3,
  sample = "sample", gene = "gene", groups = "group", cq = "cq")
}
\arguments{
\item{data}{A data frame with sample,gene and group identifier and expression values. 
Expression values can be raw Cq or on linear scale.}

\item{ctVal}{Logical, TRUE if raw Cq values}

\item{pStabLim, }{numeric value on stability threshold.}

\item{sample}{Character specifies the column name of sample identifiers (default == "sample")}

\item{gene}{Character specifies the column name of gene identifiers (default == "gene")}

\item{groups}{Logical, if TRUE, the function will consider grouping variables}

\item{cq}{Character specifies the column of cq-/expression-values}

\item{groups}{Character specifies the column name of group identifiers}
}
\description{
Analyze a data set for gene expression stability with the NormFinder algorithm
}
\author{
person("Claus", " Lindbjerg Andersen", role = "aut")
}
